Remove unnecessary base form
